Разбиваем текущий код на модули
